feat: 릴리즈 노트 세부 분리(배포된/수정중)
